refactoring all code
