+++ src/applications/app.js
@@ -1,0 +1,25 @@
+/* jshint esversion: 6 */
+
+// точка входа, с которой начинается выполнение всей модели MVC
+// import Controller from './controller/controller.js';
+// const startGame = () => new Controller();
+// startGame();
+// export default startGame;
+
+// ---
+
+// import Model from './model.js';
+// import View from './view.js';
+// import Controller from './controller.js';
+
+// initializeGame(Model, View, Controller);
+
+// import {one, two} from './model/model.js';
+// import one from './model/model.js';
+// import two from './controller/controller.js';
+
+// alert(one);
+// alert(two);
+
+import {out} from './view/view.js';
+out();
