+++ src/webpack.config.js
@@ -1,0 +1,35 @@
+// подключаем базовые переменные
+// модуль для управления путями
+const path = require('path')
+
+// сам вебпак
+// const webpack = require('webpack')
+
+module.exports = {
+  // базовый путь к проекту
+  // content: path.resolve(__dirname, 'src'),
+  
+  //указываем где будет находиться наш файл  
+  // точка входа JS
+  entry: {
+    // entry: './scripts.js',
+    // основной файл приложение
+    app: [
+      './applications/app.js'   
+    ],
+  },
+  
+  mode: "development",
+  
+	//указываем файл, куда будет компилироваться наш проект
+	output: {
+		//указываем каталог где будет собираться сборка
+		path: path.join(__dirname, ''),
+
+		//указываем как будет называться наш файл, в котором будет сборка    
+    filename: './scripts.js', //[name] - заменяет имя файла автоматом на имя, что было в основном файле (см.выше)
+
+		//Указываем относительный путь, где будет статические другие файлы
+    publicPath: 'build'
+	},
+}
